Created new applicant screen
